+++ bindings/python/ContinuousDynamicalSystem/include/BipedalLocomotion/bindings/ContinuousDynamicalSystem/Integrator.h
@@ -15,6 +15,7 @@
 #include <BipedalLocomotion/ContinuousDynamicalSystem/FixedStepIntegrator.h>
 #include <BipedalLocomotion/ContinuousDynamicalSystem/ForwardEuler.h>
 #include <BipedalLocomotion/ContinuousDynamicalSystem/Integrator.h>
+#include <BipedalLocomotion/ContinuousDynamicalSystem/RK4.h>
 
 #include <pybind11/chrono.h>
 #include <pybind11/eigen.h>
@@ -34,7 +35,7 @@ void CreateIntegrator(pybind11::module& module, const std::string& name)
     namespace py = ::pybind11;
     using namespace BipedalLocomotion::ContinuousDynamicalSystem;
 
-    const std::string completeName = "_" + name + "Integrator";
+    const std::string completeName = "_" + name + "BaseIntegrator";
     py::class_<Integrator<_Derived>, _Args...>(module, completeName.c_str())
         .def("set_dynamical_system",
              &Integrator<_Derived>::setDynamicalSystem,
@@ -82,9 +83,8 @@ void CreateFixedStepIntegrator(pybind11::module& module, const std::string& name
     namespace py = ::pybind11;
     using namespace BipedalLocomotion::ContinuousDynamicalSystem;
 
-    CreateIntegrator<FixedStepIntegrator<_Derived>>(module, name);
-
     const std::string completeName = "_" + name + "FixedStepIntegrator";
+    CreateIntegrator<FixedStepIntegrator<_Derived>>(module, name);
     py::class_<FixedStepIntegrator<_Derived>,
                Integrator<FixedStepIntegrator<_Derived>>,
                _Args...>(module, completeName.c_str())
@@ -108,11 +108,26 @@ void CreateForwardEulerIntegrator(pybind11::module& module, const std::string& n
     namespace py = ::pybind11;
     using namespace BipedalLocomotion::ContinuousDynamicalSystem;
 
-    CreateFixedStepIntegrator<ForwardEuler<_DynamicalSystem>>(module, name);
-
     const std::string completeName = name + "ForwardEulerIntegrator";
+    CreateFixedStepIntegrator<ForwardEuler<_DynamicalSystem>>(module, completeName);
+
     py::class_<ForwardEuler<_DynamicalSystem>,
                FixedStepIntegrator<ForwardEuler<_DynamicalSystem>>,
+               _Args...>(module, completeName.c_str())
+        .def(py::init());
+}
+
+template <typename _DynamicalSystem, typename... _Args>
+void CreateRK4Integrator(pybind11::module& module, const std::string& name)
+{
+    namespace py = ::pybind11;
+    using namespace BipedalLocomotion::ContinuousDynamicalSystem;
+
+    const std::string completeName = name + "RK4Integrator";
+    CreateFixedStepIntegrator<RK4<_DynamicalSystem>>(module, completeName);
+
+    py::class_<RK4<_DynamicalSystem>,
+               FixedStepIntegrator<RK4<_DynamicalSystem>>,
                _Args...>(module, completeName.c_str())
         .def(py::init());
 }